Move date format helper to more generic location
