+++ apps/nerves_hub_www/lib/nerves_hub_www_web/views/layout_view.ex
@@ -263,4 +263,14 @@ defmodule NervesHubWWWWeb.LayoutView do
       end
     end)
   end
+
+  defmodule DateTimeFormat do
+    @months %{1 => "Jan", 2 => "Feb", 3 => "Mar", 4 => "Apr",
+      5 => "May", 6 => "Jun", 7 => "Jul", 8 => "Aug",
+      9 => "Sep", 10 => "Oct", 11 => "Nov", 12 => "Dec"}
+  
+    def format(timestamp) do
+      '#{@months[timestamp.month]} #{timestamp.day}, #{timestamp.year}'
+    end
+  end
 end
